refactored vote and sawp pages
