+++ src/components/client/StakePage.tsx
@@ -6,36 +6,22 @@ import UnlockNotification from '@/components/lock/UnlockNotification'
 import {
   Flex,
   Heading,
-  Text,
-  Box,
   SimpleGrid,
-  Spacer,
-  VStack,
-  Tabs,
-  TabList,
-  TabPanels,
-  TabPanel,
-  Tab,
-  chakra,
-  Icon,
 } from '@chakra-ui/react'
 import React, { useState, useEffect, useCallback } from 'react'
-import { RiQuestionLine } from 'react-icons/ri'
 import LockOverview from '@/components/lock/LockOverview'
 import LockedDetails from '@/components/lock/LockedDetails'
 import { useLockOverview } from '@/hooks/useLockOverview'
-import * as Humanize from 'humanize-plus'
 import { useLock } from '@/hooks/useLock'
 import { useReward } from '@/hooks/useReward'
 import { useWaitForTransaction, useNetwork, useSwitchNetwork } from 'wagmi'
 import NetworkErrorNotification from '@/components/lock/NetworkErrorNotification'
 import config from '@/config'
-import StakeIQ from '@/components/lock/StakeIQ'
-import IncreaseLockTime from '@/components/lock/IncreaseLockTime'
 import { Dict } from '@chakra-ui/utils'
 import { useIQRate } from '@/hooks/useRate'
 import { useReusableToast } from '@/hooks/useToast'
 import { useLockEnd } from '@/hooks/useLockEnd'
+import { StakeBox } from '../dashboard/StakeBox'
 
 const StakePage = () => {
   const [openUnlockNotification, setOpenUnlockNotification] = useState(false)
@@ -125,19 +111,6 @@ const StakePage = () => {
     }
   }
 
-  const tabs = {
-    'Stake more IQ': {
-      label: 'Stake more IQ',
-      borderLeftRadius: '5',
-      borderRightColor: 'transparent',
-    },
-    'Increase Stake time': {
-      label: 'Increase Stake time',
-      borderRightRadius: '5',
-      borderLeftColor: 'transparent',
-    },
-  }
-
   return (
     <>
       <Flex pt={{ base: '5', lg: '6' }} direction="column" gap="6" pb="20">
@@ -155,88 +128,17 @@ const StakePage = () => {
             columns={{ base: 1, lg: 2 }}
             spacing={{ base: 8, lg: 0, '2xl': 4 }}
           >
-            <Flex
-              maxW={{ base: 526, lg: '576' }}
-              w="full"
-              p="5"
-              rounded="lg"
-              border="solid 1px"
-              borderColor="divider"
-              direction="column"
-              gap="6"
-              mx="auto"
-              mb="auto"
-            >
-              <Flex gap="2.5" align="center">
-                <Heading fontWeight="bold" fontSize={{ md: 'xl', lg: '2xl' }}>
-                  Stake IQ
-                </Heading>
-                <Spacer />
-                <Icon
-                  color="brandText"
-                  cursor="pointer"
-                  onClick={() => setOpenStakingInfo(true)}
-                  fontSize={20}
-                  as={RiQuestionLine}
-                />
-              </Flex>
-              {userTotalIQLocked > 0 && typeof lockEndDate !== 'number' && (
-                <Box
-                  alignSelf="center"
-                  rounded="md"
-                  width={{ md: 355 }}
-                  bg="lightCard"
-                  textAlign="center"
-                  p={4}
-                >
-                  <Text fontSize={{ base: 'xs', md: 'sm' }} fontWeight="medium">
-                    You have Staked a total of{' '}
-                    <chakra.span fontWeight="bold">
-                      {Humanize.formatNumber(userTotalIQLocked, 2)}
-                    </chakra.span>{' '}
-                    IQ. Expiring on{' '}
-                    <chakra.span fontWeight="bold">
-                      {lockEndDate?.toUTCString()}
-                    </chakra.span>
-                  </Text>
-                </Box>
-              )}
-              <Tabs variant="unstyled">
-                {userTotalIQLocked > 0 && (
-                  <TabList display="flex" justifyContent="center">
-                    {Object.values(tabs).map(({ label, ...tabProps }) => (
-                      <Tab
-                        key={label}
-                        px={{ base: 3, md: 4 }}
-                        border="1px solid"
-                        fontWeight={{ md: 'bold' }}
-                        fontSize="xs"
-                        borderColor="divider2"
-                        _selected={{ color: 'white', bg: 'brandText' }}
-                        {...tabProps}
-                      >
-                        {label}
-                      </Tab>
-                    ))}
-                  </TabList>
-                )}
-                <TabPanels>
-                  <TabPanel p={0} pt={6}>
-                    <StakeIQ exchangeRate={exchangeRate} />
-                  </TabPanel>
-                  <TabPanel p={0} mt={7}>
-                    <VStack rowGap={6}>
-                      <IncreaseLockTime />
-                    </VStack>
-                  </TabPanel>
-                </TabPanels>
-              </Tabs>
-            </Flex>
+            <StakeBox
+              userTotalIQLocked={userTotalIQLocked}
+              setOpenStakingInfo={setOpenStakingInfo}
+              lockEndDate={lockEndDate}
+              exchangeRate={exchangeRate}
+            />
             <LockedDetails
-              setOpenUnlockNotification={(status) =>
+              setOpenUnlockNotification={status =>
                 setOpenUnlockNotification(status)
               }
-              setOpenRewardCalculator={(status) =>
+              setOpenRewardCalculator={status =>
                 setOpenRewardCalculator(status)
               }
               loading={isProcessingUnlock}
