+++ src/types/voteType.ts
@@ -1,0 +1,10 @@
+export type VotingItemProps = {
+  item: {
+    id: string
+    title: string
+    body: string
+    author: string
+    end: number
+  }
+  active?: boolean
+}
